+++ packages/ocean-react/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.139.1](https://github.com/ocean-ds/ocean-web/compare/v1.139.0...v1.139.1) (2026-09-18)
+
+### Bug Fixes
+
+- **core:** MR-772 calendário do DateRange/DatePicker na largura dos dois campos ([#1257](https://github.com/ocean-ds/ocean-web/issues/1257)) ([0e533a5](https://github.com/ocean-ds/ocean-web/commit/0e533a55eed990e8a0bdd9f9c871934825362d3e))
+
+# [1.139.0](https://github.com/ocean-ds/ocean-web/compare/v1.138.0...v1.139.0) (2026-09-17)
+
+### Features
+
+- **drawer:** props para pilha de drawers lado a lado (MR-795) ([#1259](https://github.com/ocean-ds/ocean-web/issues/1259)) ([56d1d04](https://github.com/ocean-ds/ocean-web/commit/56d1d04aaeab16c17349dafbd95e90db49591bf1))
+
 # [1.138.0](https://github.com/ocean-ds/ocean-web/compare/v1.137.1...v1.138.0) (2026-09-11)
 
 ### Features